+++ ChangeLog.md
@@ -17,8 +17,8 @@ All notable changes to this project will be documented in this file.
 - function `io/copy-files-glob`
 - function `io/copy-file-tree`
 - function `io/move-files-glob`
-- function `io/file-create-symbolic-link`
-- function `io/file-create-hard-link`
+- function `io/create-symbolic-link`
+- function `io/create-hard-link`